TreeNode extensions documentation / bump SimDB (#644)
I moved some files around without breaking backwards compatibility. I
did this so I can have a README file in the extensions directory. I can
also reformat the markdown and move this instead to `sparta_docs` if
that's where it should be -- it was just easier to write a markdown
file.

Whether we use a README or a Sparta doc, I still like having all the
extensions code in one spot. I can refactor further so we only have
`TreeNodeExtensions.cpp|hpp`

The new code in SimDB is needed for CherryPickFastCheckpointer and
CoSimEventPipeline (more generally, the changes are needed for any use
case that will instantiate multiple instances of the same app). The
previous code required that you have an AppManager before configuring
the factories, but the intended use for Sparta simulators is that you go
through CommandLineSimulator/SimulationConfiguration to set up the apps,
then let the Simulation class actually create them. This wasn't caught
before since the CherryPickFastCheckpointer is only used in a unit test
in Sparta (no simulator). And it wasn't caught in PegasusCoSim since
that simulator manually creates the DatabaseManager / AppManager
(discouraged - there will be a follow up PR to do things the right way
in Pegasus).
