+++ sparta/src/TreeNodeExtensions.cpp
@@ -17,7 +17,6 @@
 #include "sparta/utils/KeyValue.hpp"
 #include "sparta/utils/SpartaAssert.hpp"
 #include "sparta/utils/SpartaException.hpp"
-#include "sparta/utils/Utils.hpp"
 
 namespace YAML {
 class EventHandler;
@@ -228,18 +227,6 @@ const std::string & ExtensionDescriptor::getName() const
 std::unique_ptr<ParameterSet> ExtensionDescriptor::cloneParameters() const
 {
     return impl_->cloneParameters();
-}
-
-/*!
- * \brief TreeNode::ExtensionsBase class implementation
- */
-TreeNode::ExtensionsBase::ExtensionsBase()
-    : uuid_(generateUUID())
-{
-}
-
-TreeNode::ExtensionsBase::~ExtensionsBase()
-{
 }
 
 /*!
